+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## unreleased
 
 - Fixed: cannot convert to OGG with an external cover art
+- Fixed: track number detection can't parse X/Y format
 - Removed: support for any output format except MP3 and OGG