+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## unreleased
 
+- Added: APE as one of default input formats
 - Fixed: cannot convert to OGG with an external cover art
 - Fixed: track number detection can't parse X/Y format
 - Fixed: CUE "performer" is not treated as "artist"